requirements.txt
plotly-express==0.4.1
requests==2.32.3
streamlit==1.36.0
